Accessing and Updating Bookcode Information
Just as there can be error's in the songs data there can also be errors in the bookcodes data and we want casual users to be able to correct that as well. If you examine the fakebooks3005winter2016.db you will notice it has a bookcodes tables that the fakebooks3005fall2018.db database does not. This table provides data about the books that the bookcode in the songs table represent.

Taking the bookcodes table from the 2016 database and add it to the 2018 database using in the previous problem. Then modify the application so that casual users can edit the book data as well. Specifically, repeat the process from the previous problem but now allow the user to search for and correct information in the bookcodes table as well. Modify the code to also satisfy the following requirements.